+++ CHANGELOG.md
@@ -7,6 +7,11 @@ fájl a lényegi, ember által írt kiemeléseket rögzíti.
 
 ## [Nem kiadott]
 
+### Hozzáadva
+- **A személy-album fejlécén az arc ↔ teljes kép váltóval a képek a
+  személy arcára közelítve is megjeleníthetők (#2187).** A közelített kép
+  éles, kis arcnál is, mert a program ilyenkor nagyobb felbontásból vág.
+
 ## [0.8.583] – 2026-09-26
 
 ### Javítva